requirements.txt
Flask==2.3.3
Biopython==1.81
Werkzeug==2.3.7
numpy==1.26.4
pandas==2.2.1 